moved firebase_config.js to src/firebase_config.js
