+++ src/app/ai-conversations/page.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Toolbar from '../../components/Toolbar';
 import { useState, useEffect } from 'react';
 import { useAuth } from '../context/AuthContext';
-import { firebaseDB } from '../../../firebase_config';
+import { firebaseDB } from '../../firebase_config';
 import { getDatabase, ref, get, set } from 'firebase/database';
 
 export default function GeminiChatbot() {
