+++ src/app/quizzes/page.js
@@ -4,7 +4,7 @@ import React, { useState, useEffect } from 'react';
 import Toolbar from '../../components/Toolbar';
 import Link from 'next/link';
 import { useAuth } from '../context/AuthContext';
-import { firebaseDB } from '../../../firebase_config';
+import { firebaseDB } from '../../firebase_config';
 import { getDatabase, ref, get, set } from 'firebase/database';
 export default function Quizzes() {
   const [quiz, setQuiz] = useState(null);
